+++ react/__mocks__/vtex.native-types/index.js
@@ -2,3 +2,4 @@ import React from 'react'
 import { FormattedMessage } from 'react-intl'
 
 export const IOMessage = props => <FormattedMessage {...props} />
+export const formatIOMessage = message => message